refactor: share mixin and types across Select and FilterMultiSelect
fix: lint
