+++ packages/select/index.ts
@@ -1,2 +1,3 @@
 export * from "./src/Select"
 export * from "./src/FilterMultiSelect"
+export * from "./src/types"
