+++ packages/select/src/FilterMultiSelect/components/ListBoxSection/ListBoxSection.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { Node } from "@react-types/shared"
-import { ItemType } from "../../types"
+import { ItemType } from "../../../types"
 import styles from "./ListBoxSection.module.scss"
 
 export interface ListBoxSectionProps {
