+++ packages/select/src/FilterMultiSelect/components/SelectionControlButton/SelectionControlButton.module.scss
@@ -2,7 +2,7 @@
 @import "~@kaizen/design-tokens/sass/color";
 @import "~@kaizen/design-tokens/sass/typography";
 @import "~@kaizen/design-tokens/sass/spacing";
-@import "../../mixins";
+@import "../../../mixins";
 
 $focus-ring-offset: 1px;
