+++ packages/select/src/FilterMultiSelect/index.ts
@@ -3,5 +3,4 @@ export * from "./FilterMultiSelect"
 // Here only export component props as we would like to enforce consumer to use dot notation component.
 export * from "./components"
 export * from "./provider"
-export * from "./types"
 export { getSelectedOptionLabels, getSelectedOptionKeys } from "./utils"
