+++ packages/select/src/FilterMultiSelect/utils/getSelectedOptionKeys/getSelectedOptionKeys.ts
@@ -1,6 +1,6 @@
 import React from "react"
 import { Selection } from "@react-types/shared"
-import { ItemType } from "../../types"
+import { ItemType } from "../../../types"
 
 export const getSelectedOptionKeys = (
   keys?: Selection,
