+++ packages/select/src/FilterMultiSelect/utils/getSelectedOptionLabels/getSelectedOptionLabels.spec.ts
@@ -1,4 +1,4 @@
-import { ItemType } from "../../types"
+import { ItemType } from "../../../types"
 import { getSelectedOptionLabels } from "./getSelectedOptionLabels"
 
 const itemsMock: ItemType[] = [
