--- packages/select/src/Select/_mixins.scss
@@ -1,9 +1,0 @@
-// reset user agent styles for button element type
-@mixin button-reset {
-  appearance: none;
-  background: transparent;
-  border: none;
-  font: inherit;
-  margin: 0;
-  transition: none; // override Murmur global styles :(
-}
