+++ packages/select/src/Select/components/index.ts
@@ -1,2 +1,3 @@
 export type { TriggerButtonProps } from "./TriggerButton"
 export type { OptionProps } from "./Option"
+export type { SingleListBoxProps } from "./ListBox"
