+++ packages/select/src/Select/index.ts
@@ -2,4 +2,3 @@ export * from "./Select"
 
 // Here only export component props as we would like to enforce consumer to use dot notation component.
 export * from "./components"
-export * from "./types"
