--- packages/select/src/Select/types.ts
@@ -1,11 +1,0 @@
-import { SelectState } from "@react-stately/select"
-
-export type ValueType = React.Key
-export interface ItemType {
-  label: string
-  value: ValueType
-}
-
-export type State = {
-  state: SelectState<ItemType>
-}
